requirements.txt
importlib_resources
dataclasses
bscscan-python
python-binance
selenium
pyTelegramBotAPI
web3